TG-40: new folder structure for list and form components
