+++ src/Components/Shared/Form/Form.jsx
@@ -1,47 +1,35 @@
 import { useState, useEffect } from 'react';
-import { useLocation, useParams, useHistory, withRouter } from 'react-router-dom';
+import { useHistory, withRouter } from 'react-router-dom';
 import style from './styles.module.css';
 import Button from '../Button/Button';
 
-const Form = ({ data, dbPath }) => {
-  const { linkData, itemData, DBPath } = useLocation();
-  const { id } = useParams();
+const Form = ({ data, itemData, submitHandler }) => {
   const { goBack } = useHistory();
   const [inputValues, setInputValues] = useState({});
-  const [config, setConfig] = useState([]);
-  const url = dbPath || DBPath;
 
   // === Create instance state on mount === //
   useEffect(() => {
     let template = {};
     if (data) {
-      setConfig(data);
-      data.forEach((item) => {
-        if (item.type === 'checkbox') template[item.key] = false;
-        else template[item.key] = '';
-      });
-      setInputValues(template);
-    } else if (linkData) {
-      setConfig(linkData);
       if (itemData) {
         let formattedItem = {};
-        linkData.forEach((item) => {
+        data.forEach((item) => {
           if (itemData[item.key] && typeof itemData[item.key] === 'object') {
             formattedItem[item.key] = itemData[item.key]._id;
           } else if (item.type === 'date') {
-            formattedItem[item.key] = itemData[item.key].substring(0, 10);
+            formattedItem[item.key] = itemData[item.key];
           } else formattedItem[item.key] = itemData[item.key];
         });
         setInputValues(formattedItem);
       } else {
-        linkData.forEach((item) => {
+        data.forEach((item) => {
           if (item.type === 'checkbox') template[item.key] = false;
           else template[item.key] = '';
         });
         setInputValues(template);
       }
     }
-  }, []);
+  }, [itemData]);
 
   // === Handle value change for different input types === //
   const handleChange = (e, input) => {
@@ -52,55 +40,9 @@ const Form = ({ data, dbPath }) => {
     else setInputValues({ ...inputValues, [input.key]: e.target.value });
   };
 
-  // === Fetch functions === key
-  const createInstance = async (obj) => {
-    try {
-      const res = await fetch(`${process.env.REACT_APP_API_URL}${url}`, {
-        method: 'POST',
-        headers: { 'content-type': 'application/json' },
-        body: JSON.stringify(obj)
-      });
-      const body = await res.json();
-      return { message: body.message, err: body.error };
-    } catch (error) {
-      alert(error);
-    }
-  };
-
-  const updateInstance = async (obj) => {
-    try {
-      const res = await fetch(`${process.env.REACT_APP_API_URL}${url}/${id}`, {
-        method: 'PUT',
-        headers: { 'content-type': 'application/json' },
-        body: JSON.stringify(obj)
-      });
-      const body = await res.json();
-      return { message: body.message, err: body.error };
-    } catch (error) {
-      alert(error);
-    }
-  };
-
-  // === Handle submit data and method === //
-  const submitHandler = async (e) => {
-    e.preventDefault();
-    let result;
-
-    if (id) {
-      result = await updateInstance(inputValues);
-    } else {
-      result = await createInstance(inputValues);
-    }
-
-    if (result && result.error === false) setInputValues({});
-    alert(result.message);
-
-    if (id) goBack();
-  };
-
   return (
     <form className={style.form} onSubmit={submitHandler}>
-      {config.map((item) => {
+      {data.map((item) => {
         return (
           <div
             key={item.key}
