+++ src/Components/Tasks/List/tasks.module.css
@@ -1,0 +1,91 @@
+.container {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.btn {
+  display: inline-block;
+  background: #EA3354;
+  color: #000;
+  border: none;
+  padding: 10px 20px;
+  margin: 15px;
+  border-radius: 5px;
+  cursor: pointer;
+  text-decoration: none;
+  font-size: 15px;
+  font-family: inherit;
+}
+
+.btn:focus {
+  outline: none;
+}
+
+.btn:active {
+  transform: scale(0.98);
+}
+
+.btnBlock {
+  display: block;
+  width: 98%;
+}
+
+.taskForm {
+  margin-bottom: 40px;
+  border: 3px solid black;
+  border-radius: 10px;
+  width: 80%;
+}
+
+.readOnly{
+  color: rgba(0,0,0,0.5);
+}
+
+.inputContainer {
+  margin: 20px;
+}
+
+.inputContainer label {
+  display: block;
+}
+
+.inputContainer input {
+  width: 98%;
+  height: 25px;
+  margin: 5px;
+  padding: 3px 7px;
+  font-size: 17px;
+}
+
+.inputContainer select {
+  width: 100%;
+  height: 35px;
+  margin: 5px;
+  padding: 3px 7px;
+  font-size: 17px;
+}
+
+.check {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: max-content;
+}
+
+.check label {
+  flex: 1;
+  cursor: pointer;
+}
+
+.check input {
+  flex: 2;
+  height: 20px;
+  cursor: pointer;
+}
+
+.data {
+  margin: 10px;
+}
