--- src/Components/Tasks/ListItem/listItem.component.css
@@ -1,4 +1,0 @@
-.rows{
-    background-color: gray;
-    border: 1px solid black;
-}
